# Week7/generated/openapi.yaml
openapi: 3.0.3
info:
  description: API for managing products.
  title: Product API
  version: 1.0.0
servers:
- url: http://localhost:8000
paths:
  /products:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_products_get_200_response"
          description: Products retrieved
          headers:
            Cache-Control:
              description: Cache directive for the response
              explode: false
              schema:
                type: string
              style: simple
      summary: List products
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_products_post_201_response"
          description: Product created
          headers:
            Cache-Control:
              description: Cache directive for the response
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Invalid payload
      summary: Create product
  /products/{productId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: productId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          description: Product deleted
          headers:
            Cache-Control:
              description: Cache directive for the response
              explode: false
              schema:
                type: string
              style: simple
        "404":
          description: Product not found
      summary: Delete product
    get:
      parameters:
      - explode: false
        in: path
        name: productId
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_products_post_201_response"
          description: Product retrieved
          headers:
            Cache-Control:
              description: Cache directive for the response
              explode: false
              schema:
                type: string
              style: simple
        "404":
          description: Product not found
      summary: Get product
    put:
      parameters:
      - explode: false
        in: path
        name: productId
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_products_post_201_response"
          description: Product updated
          headers:
            Cache-Control:
              description: Cache directive for the response
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: Invalid payload
        "404":
          description: Product not found
      summary: Replace product
components:
  schemas:
    Product:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        price: 6.0274563
        name: name
        description: description
        id: 0
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          format: int64
          title: id
          type: integer
        name:
          title: name
          type: string
        price:
          format: float
          title: price
          type: number
        description:
          nullable: true
          title: description
          type: string
        createdAt:
          format: date-time
          title: createdAt
          type: string
        updatedAt:
          format: date-time
          title: updatedAt
          type: string
      required:
      - createdAt
      - id
      - name
      - price
      - updatedAt
      title: Product
      type: object
    ProductInput:
      example:
        price: 0.8008282
        name: name
        description: description
      properties:
        name:
          title: name
          type: string
        price:
          format: float
          title: price
          type: number
        description:
          nullable: true
          title: description
          type: string
      required:
      - name
      - price
      title: ProductInput
      type: object
    _products_get_200_response:
      example:
        data:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          price: 6.0274563
          name: name
          description: description
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
        - createdAt: 2000-01-23T04:56:07.000+00:00
          price: 6.0274563
          name: name
          description: description
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        data:
          items:
            $ref: "#/components/schemas/Product"
          title: data
          type: array
      title: _products_get_200_response
      type: object
    _products_post_201_response:
      example:
        data:
          createdAt: 2000-01-23T04:56:07.000+00:00
          price: 6.0274563
          name: name
          description: description
          id: 0
          updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        data:
          $ref: "#/components/schemas/Product"
      title: _products_post_201_response
      type: object